Factored out compiler module
