+++ panel/io/server.py
@@ -4,6 +4,7 @@ Utilities for creating bokeh Server instances.
 from __future__ import absolute_import, division, unicode_literals
 
 import signal
+import threading
 
 from functools import partial
 
@@ -100,3 +101,27 @@ def get_server(panel, port=0, websocket_origin=None, loop=None,
         except RuntimeError:
             pass
     return server
+
+
+class StoppableThread(threading.Thread):
+    """Thread class with a stop() method."""
+
+    def __init__(self, io_loop=None, timeout=1000, **kwargs):
+        from tornado import ioloop
+        super(StoppableThread, self).__init__(**kwargs)
+        self._stop_event = threading.Event()
+        self.io_loop = io_loop
+        self._cb = ioloop.PeriodicCallback(self._check_stopped, timeout)
+        self._cb.start()
+
+    def _check_stopped(self):
+        if self.stopped:
+            self._cb.stop()
+            self.io_loop.stop()
+
+    def stop(self):
+        self._stop_event.set()
+
+    @property
+    def stopped(self):
+        return self._stop_event.is_set()
