+++ panel/models/vega.py
@@ -6,7 +6,8 @@ import os
 from bokeh.core.properties import Dict, String, Any, Instance
 from bokeh.models import LayoutDOM, ColumnDataSource
 
-from ..util import CUSTOM_MODELS
+from ..compiler import CUSTOM_MODELS
+
 
 class VegaPlot(LayoutDOM):
     """
